AARAMBH -Added CriteriaBuilder SpecificationExecutor for JPA Repository in ReportsService.java.

+++ src/main/java/com/oms/dto/requests/ProductShipmentManager.java
@@ -10,7 +10,7 @@ public class ProductShipmentManager {
     private Long scheduleQty;
     private Long pendingQty;
     private Long suppliedQty;
-    private float pov;
+    private double pov;
     private Long productOrderId;
     private String esplPO;
     private String invoiceNo;
